pranav/pyproject.toml
[project]
name = "pranav"
version = "0.1.0"
description = "Pranav AI - RAG Chatbot Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.0",
    "openai>=1.0.0",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.0.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
